Including an updated version of perl.vim syntax highlighting as the one provided with mac os x does a pretty poor job of handling some expressions.
I just cherry picked this out of http://github.com/petdance/vim-perl, which is
the git repository for the official maintainer of syntax and such for vim
related to perl.  There is a lot more stuff in the package, including perl6
stuff, but I don't need any of that and didn't want to clutter up my
environment.